add styling to widget list
