+++ components/DashboardWidgets/UpcomingWidget/UpcomingWidget.styled.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const StyledUnorderedList = styled.ul`
+  display: flex;
+  flex-direction: column;
+  gap: 5px;
+`;
